+++ src/main/java/com/readum/domain/user/userbook/dto/UserBookCreateResult.java
@@ -1,4 +1,4 @@
-package com.readum.domain.userbook.dto;
+package com.readum.domain.user.userbook.dto;
 
 import java.time.LocalDateTime;
