+++ src/main/java/com/readum/model/user/repository/UserBookRepository.java
@@ -1,6 +1,6 @@
-package com.readum.model.book.repository;
+package com.readum.model.user.repository;
 
-import com.readum.model.book.entity.UserBook;
+import com.readum.model.user.entity.UserBook;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
