+++ src/main/java/com/readum/presentation/controller/user/dto/UserBookCreateRequest.java
@@ -1,6 +1,6 @@
-package com.readum.presentation.controller.userbook.dto;
+package com.readum.presentation.controller.user.dto;
 
-import com.readum.domain.userbook.dto.UserBookCreateCommand;
+import com.readum.domain.user.userbook.dto.UserBookCreateCommand;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
