+++ src/test/java/com/readum/domain/aiChat/service/AiChatSessionCreateServiceTest.java
@@ -6,8 +6,8 @@ import com.readum.domain.aiChat.exception.AiChatErrorCode;
 import com.readum.domain.exception.NotFoundException;
 import com.readum.model.aiChat.entity.AiChatSession;
 import com.readum.model.aiChat.repository.AiChatSessionRepository;
-import com.readum.model.book.entity.UserBook;
-import com.readum.model.book.repository.UserBookRepository;
+import com.readum.model.user.entity.UserBook;
+import com.readum.model.user.repository.UserBookRepository;
 import org.assertj.core.api.InstanceOfAssertFactories;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
